+++ packages/ui/src/app/pages/WorkingGroups/WorkingGroup/WorkingGroup.stories.tsx
@@ -320,7 +320,9 @@ export const CreateOpening: Story = {
     })
 
     step('Transaction parameters', () => {
-      const [description, openingType, stakePolicy, rewardPerBlock] = args.onCreateOpening.mock.calls.at(-1)
+      const [signer, description, openingType, stakePolicy, rewardPerBlock] = args.onCreateOpening.mock.calls.at(-1)
+
+      expect(signer).toBe(member('alice').controllerAccount)
 
       expect(stakePolicy.toJSON()).toEqual({
         stakeAmount: 100_0000000000,
@@ -397,7 +399,9 @@ export const CreateOpeningImport: Story = {
     })
 
     step('Transaction parameters', () => {
-      const [description, openingType, stakePolicy, rewardPerBlock] = args.onCreateOpening.mock.calls.at(-1)
+      const [signer, description, openingType, stakePolicy, rewardPerBlock] = args.onCreateOpening.mock.calls.at(-1)
+
+      expect(signer).toBe(member('alice').controllerAccount)
 
       expect(stakePolicy.toJSON()).toEqual({
         stakeAmount: 200_0000000000,