+++ packages/ui/src/common/components/Warning.tsx
@@ -32,10 +32,10 @@ export const Warning = ({ title, content, isClosable, additionalContent, icon, i
         <HeaderWrapper>
           {icon === 'alert' && <AlertSymbol />}
           {icon === 'info' && <InfoSymbol />}
-          {title && <h5>{title}</h5>}
+          {title ? <h5>{title}</h5> : content}
         </HeaderWrapper>
       )}
-      {content && (
+      {title && content && (
         <TextMedium as="div" inter light>
           {content}
         </TextMedium>