+++ packages/ui/src/memberships/components/MemberProfile/MemberDetails.tsx
@@ -13,6 +13,8 @@ import {
   SidePaneLabel,
   EmptyBody,
 } from '@/common/components/SidePane'
+import { has } from '@/common/utils/object'
+import { ExternalResourceLink } from '@/memberships/constants'
 import { useIsMyMembership } from '@/memberships/hooks/useIsMyMembership'
 import { useMemberExtraInfo } from '@/memberships/hooks/useMemberExtraInfo'
 
@@ -36,15 +38,6 @@ export const MemberDetails = React.memo(({ member }: Props) => {
     councilMember = '-',
     initiatingLeaving = '-',
   } = useMemberExtraInfo(member)
-
-  const externalResourceLink: any = {
-    TELEGRAM: 'https://web.telegram.org/k/#@',
-    TWITTER: 'https://twitter.com/',
-    FACEBOOK: 'https://facebook.com/',
-    YOUTUBE: 'https://youtube.com/user/',
-    LINKEDIN: 'https://www.linkedin.com/in/',
-    GITHUB: 'https://github.com/',
-  }
 
   if (isLoading || !memberDetails) {
     return (
@@ -143,9 +136,9 @@ export const MemberDetails = React.memo(({ member }: Props) => {
           <SidePaneRow key={`${externalResource.source}-externalResources`}>
             <SidePaneLabel text={socialTitle(externalResource.source)} />
             <SidePaneText>
-              {externalResourceLink[externalResource.source] ? (
+              {has(externalResource.source, ExternalResourceLink) ? (
                 <Link
-                  href={`${externalResourceLink[externalResource.source]}${externalResource.value}`}
+                  href={`${ExternalResourceLink[externalResource.source]}${externalResource.value}`}
                   target="_blank"
                 >
                   {externalResource.value}