+++ packages/ui/src/validators/types/index.ts
@@ -1,4 +1,1 @@
 export * from './Validator'
-
-export type Verification = null | 'verified' | 'unverified'
-export type State = null | 'active' | 'waiting'